requirements.txt
ccxt==4.3.89
pandas==2.2.2
ta==0.11.0
python-dotenv==1.0.1
python-telegram-bot==21.3
APScheduler==3.10.4
sqlalchemy==2.0.30
requests==2.32.3
pytest==8.2.2
pytest-asyncio==0.23.7
numpy==1.26.4
fastapi==0.111.0
uvicorn==0.30.1
aiosqlite==0.20.0
aiofiles==23.2.1
